+++ src/petaly/utils/file_handler.py
@@ -214,7 +214,7 @@ class FileHandler:
         if not os.path.isdir(path_to_dir):
             os.makedirs(path_to_dir)
 
-    def get_all_dir_files(self, dir_path, file_extension, file_names_only=False):
+    def deprecated_get_all_dir_files(self, dir_path, file_extension, file_names_only=False):
         """
         """
 
@@ -227,6 +227,22 @@ class FileHandler:
                     result_arr.append(os.path.join(dir_path,files))
                 else:
                     result_arr.append(files)
+            else:
+                continue
+
+        return result_arr
+
+    def get_file_names_with_extensions(self, dir_path, *file_extensions):
+        """
+        """
+
+        # iterating over all files with determine extension
+        result_arr = []
+
+        for files in os.listdir(dir_path):
+            for file_extension in file_extensions:
+                if files.endswith(file_extension):
+                        result_arr.append(files)
             else:
                 continue
 
@@ -300,10 +316,10 @@ class FileHandler:
 
         logger.debug(f"The file {base_fname} has been backed up to the: {os.path.join(target_dir,target_file)}")
 
-    def cleanup_files(self, path_to_dir, file_extension):
+    def deprecated_cleanup_files(self, path_to_dir, file_extension):
         """ This function remove all files in a folder of path_to_dir from specific extension.
         """
-        files = self.get_all_dir_files(path_to_dir, file_extension)
+        files = self.deprecated_get_all_dir_files(path_to_dir, file_extension)
         for file in files:
             os.remove(file)
         return True